% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.model.R
\name{build.model.crossval}
\alias{build.model.crossval}
\title{Train model Random Forest}
\usage{
build.model.crossval(x, y, list.selected.var, list.index.cross, nvar)
}
\arguments{
\item{x}{input data where columns are variables and rows are observations (all numeric)}

\item{y}{decision variable as a boolean vector of length equal to number of observations}

\item{list.selected.var}{A \code{\link{list}} with selected variables in cross-validation}

\item{list.index.cross}{A \code{\link{list}} with indexes obtained in cross-validation}

\item{nvar}{the number of first variables for which to train model Random Forest}
}
\value{
A \code{\link{list}} with metrics Accuracy, AUC, MCC
}
\description{
Train model Random Forest
}
\details{
Train model Random Forest in cross-validation and variables selected in the cross-validatio
}
\examples{
\dontrun{

class <- data$class
data$class <- NULL

list.index.cross <- cross.val(x = data,
                             y = decisions,
                             method = 'kfoldcv',
                             params.cv = list(niter = 10, k = 3))

list.selected.var <- feature.selection(x = data,
                             y = class,
                             method = 'fs.utest',
                             list.index.cross = list.index.cross,
                             params = list(adjust = 'holm', alpha = 0.05))

model.result <- build.model.crossval(x = data,
                                     y = class,
                                     list.selected.var = list.selected.var,
                                     list.index.cross = list.index.cross,
                                     nvar = 10)

}

}
