% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corelation.R
\name{corelation.search}
\alias{corelation.search}
\title{Function to search highly correlated variables}
\usage{
corelation.search(data, level.cor)
}
\arguments{
\item{data}{input data where columns are variables and rows are observations (all numeric)}

\item{level.cor}{cutoff level of correlated variables. If equal to 1 is not performed}
}
\value{
A \code{\link{vector}} with indexes of highly correlated variables
}
\description{
Function to search highly correlated variables
}
\details{
Searching highly  correlated variables
}
