% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.validation.R
\name{cross.val}
\alias{cross.val}
\title{Cross-validation}
\usage{
cross.val(x, y, method, params.cv = list(niter = 10, k = 3, test.size = 0.3))
}
\arguments{
\item{x}{input data where columns are variables and rows are observations (all numeric)}

\item{y}{decision variable as a boolean vector of length equal to number of observations}

\item{method}{cross-validation method \code{kfoldcv} for cross-validation \code{k-fold} or \code{rsampling} for \code{random sampling}}

\item{params.cv}{A \code{\link{list}} with the following fields:
\itemize{
\item \code{niter} -- the number of validation repetitions
\item \code{k} -- the number of groups that a given data sample is to be split into
\item \code{test.size} -- testing set size for random sampling validation
}}
}
\value{
A \code{\link{list}} with lindices divided into train and test subsamples niter repetitions.
}
\description{
Cross-validation
}
\details{
creates a list with division of observational indices of the dataframe of
observational indices into training and test subsamples of a given number of iterations
Implements two methods : cross-validation k-fold and random sampling
}
\examples{
\dontrun{

decisions <- data$class
data$class <- NULL

list.index.cross <- cross.val(x = data,
                             y = decisions,
                             method = 'kfoldcv',
                             params.cv = list(niter = 10, k = 3))
}

}
