% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.selection.R
\name{feature.selection.cv}
\alias{feature.selection.cv}
\title{Selection of features in cross-validation}
\usage{
feature.selection.cv(
  x,
  y,
  method,
  list.index.cross,
  params = list(adjust = "holm", alpha = 0.05)
)
}
\arguments{
\item{x}{input data where columns are variables and rows are observations (all numeric)}

\item{y}{decision variable as a boolean vector of length equal to number of observations}

\item{method}{the name of the selection method for traits available in this library}

\item{list.index.cross}{A \code{\link{list}} with indexes obtained in cross-validation}

\item{params}{A \code{\link{list}} parameters required for this method of feature selection}
}
\value{
A \code{\link{list}} with selected features for each iteration of cross-validation
}
\description{
Selection of features in cross-validation
}
\details{
Uses the methods of selection of features  available in this library
}
\examples{

\dontrun{

class <- data$class
data$class <- NULL

list.index.cross <- cross.val(x = data,
                              y = class,
                              method= 'kfoldcv',
                              params.cv = list(k = 3, niter = 10))

list.selected.var <- feature.selection.cv(x = data,
                             y = class,
                             method = 'fs.utest',
                             list.index.cross = list.index.cross,
                             params = list(adjust = 'holm', alpha = 0.05))

}
}
