% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.selection.R
\name{feature.selection}
\alias{feature.selection}
\title{Selection of features}
\usage{
feature.selection(
  x,
  y,
  methods,
  connection.method = "sum",
  params = list(adjust = "holm", feature.number = 10, alpha = 0.05)
)
}
\arguments{
\item{x}{input data where columns are variables and rows are observations (all numeric)}

\item{y}{decision variable as a boolean vector of length equal to number of observations}

\item{methods}{the name of the selection method for traits available in this library}

\item{connection.method}{method of combining selection results if more than one method is selected, options \code{'sum'} for the sum of all selected filtration methods or \code{'intersect'} for part of the common}

\item{params}{A \code{\link{list}} parameters required for this method of feature selection}
}
\value{
A  new \code{\link{data.frame}} consisting of the selected functions and the decision variable
}
\description{
Selection of features
}
\details{
Uses the methods of selection of features  available in this library
}
\examples{

\dontrun{

class <- data$class
data$class <- NULL

feature.selection(x = data,
                  y = class,
                  methods = c('fs.utest','fs.mdfs.1D', 'fs.mrmr', 'fs.mcfs'),
                  connection.method = 'sum',
                  params = list(feature.number=100, adjust = 'holm', alpha = 0.05))
}
}
