% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.mdfs.1D.R
\name{fs.mdfs.1D}
\alias{fs.mdfs.1D}
\title{Build MultiDimensional Feature Selector from IGs}
\usage{
fs.mdfs.1D(x, y, params = list(adjust = "holm", alpha = 0.05))
}
\arguments{
\item{x}{input data where columns are variables and rows are observations (all numeric)}

\item{y}{decision variable as a boolean vector of length equal to number of observations}

\item{params}{method as accepted by \code{\link[stats]{p.adjust}} (\code{"BY"}
is recommended for FDR, see Details)}
}
\value{
A \code{\link{data.frame}} with selected features and p.value
}
\description{
Build MultiDimensional Feature Selector from IGs
}
\details{
Build MultiDimensional Feature Selector from IGs
}
\examples{
\dontrun{

decisions <- data$class
data$class <- NULL

fs.mdfs.1D(data, decisions, params = list(adjust = 'holm', alpha = 0.05))
}


}
