% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.mrmr.R
\name{fs.mrmr}
\alias{fs.mrmr}
\title{Minimum Redundancy Maximal Relevancy}
\usage{
fs.mrmr(x, y, params = list(feature.number = 100))
}
\arguments{
\item{x}{input data where columns are variables and rows are observations (all numeric)}

\item{y}{decision variable as a boolean vector of length equal to number of observations}

\item{params}{number of attributes to select. Must not exceed, feature.number there cannot be more than the number of columns}
}
\value{
A \code{\link{data.frame}} with selected features
}
\description{
Minimum Redundancy Maximal Relevancy
}
\details{
Performs MRMR for feature selection
}
\examples{

\dontrun{

decisions <- data$class
data$class <- NULL

fs.mrmr(data, decisions, params = list(feature.number = 100))
}

}
