% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.utest.R
\name{fs.utest}
\alias{fs.utest}
\title{Test U Manna-Whitneya (U-test) for feature selection}
\usage{
fs.utest(x, y, params = list(adjust = "holm", alpha = 0.05))
}
\arguments{
\item{x}{input data where columns are variables and rows are observations (all numeric)}

\item{y}{decision variable as a boolean vector of length equal to number of observations}

\item{params}{method as accepted by \code{\link[stats]{p.adjust}} (\code{"BY"}
is recommended for FDR, see Details)}
}
\value{
A \code{\link{data.frame}} with selected features and p.value
}
\description{
Test U Manna-Whitneya (U-test) for feature selection
}
\details{
In case of FDR control it is recommended to use Benjamini-Hochberg-Yekutieli
p-value adjustment
}
\examples{

\dontrun{

decisions <- data$class
data$class <- NULL

fs.utest(data, decisions, params = list(adjust = 'holm', alpha = 0.05))
}

}
