% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.information.R
\name{gene.info.top.gene}
\alias{gene.info.top.gene}
\title{Getting information about top genes from gprogiler2}
\usage{
gene.info.top.gene(gene.top, condition.methods = "union")
}
\arguments{
\item{gene.top}{A \code{\link{list}} with top genes for each method specified in \code{ensembleFS}}

\item{condition.methods}{A way to combine the best genes for further analysis obtained by different methods in \code{ensembleFS}. For part of the common \code{intersect}, for union from methods \code{union}.}
}
\value{
A \code{\link{data.frame}} with gene information
}
\description{
Getting information about top genes from gprogiler2
}
\details{
Getting information about top genes from gprofiler2 obtained from ensembleFS.
Requires internet connection.
}
