% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.information.R
\name{get.info.gprofiler}
\alias{get.info.gprofiler}
\title{Getting information about genes from gprogiler2}
\usage{
get.info.gprofiler(rel.var)
}
\arguments{
\item{rel.var}{A \code{\link{vector}} with gene names}
}
\value{
A \code{\link{data.frame}} with gene information
}
\description{
Getting information about genes from gprogiler2
}
\details{
Getting information about genes from gprogiler2, requires internet connection
}
\examples{
\dontrun{

gene <- c('TMC8')
get.info.gprofiler(gene)
}

}
