% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.information.R
\name{get.info.top.gene}
\alias{get.info.top.gene}
\title{Getting information about top genes from gprogiler2}
\usage{
get.info.top.gene(gene.top, condition.methods = "union")
}
\arguments{
\item{gene.top}{A \code{\link{list}} with top genes for each method specified in \code{ensembleFS}}

\item{condition.methods}{A way to combine the best genes for further analysis obtained by different methods in \code{ensembleFS}. For part of the common \code{intersect}, for union from methods \code{union}.}
}
\value{
A \code{\link{data.frame}} with gene information
}
\description{
Getting information about top genes from gprogiler2
}
\details{
Getting information about top genes from gprofiler2 obtained from ensembleFS.
Requires internet connection.
}
\examples{
\dontrun{

decisions <- data$class
data$class <- NULL

result <- ensembleFS(data,
            decisions,
            methods = c('fs.utest', 'fs.mrmr'),
            method.cv = 'kfoldcv',
            params.cv = list(k = 3, iter = 10),
            level.cor = 0.75,
            params = list(adjust = 'fdr', feature.number = 10, alpha = 0.05),
            asm = c('fs.utest', 'fs.mrmr'),
            model = c('fs.utest', 'fs.mrmr')
            )

gene.top <- get.top.gene(result$selected.feature, 15 , 20)
info.gene <- gene.info.top.gene(gene.top, condition.methods = 'union')

}

}
