% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.information.R
\name{get.top.gene}
\alias{get.top.gene}
\title{Get the best genes}
\usage{
get.top.gene(list.imp.var.cv, level.freq, number.gene)
}
\arguments{
\item{list.imp.var.cv}{the main genes obtained in \code{ensembleFS}}

\item{level.freq}{how many times a gene has occurred in of the feature subsets}

\item{number.gene}{top number of main genes}
}
\value{
A \code{\link{list}} with top genes for each method specified in \code{ensembleFS}
}
\description{
Get the best genes
}
\details{
Get the best genes out obtained in \code{ensembleFS}
}
