% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.result.R
\name{graph.result}
\alias{graph.result}
\title{Showing result ensemble}
\usage{
graph.result(x, y)
}
\arguments{
\item{x}{A \code{\link{data.frame}} with data result metrics model and  Lustgarten’s stability measure}

\item{y}{what is showing , please selected : stability, acc, auc, mcc.}
}
\value{
graph showing
}
\description{
Showing result ensemble
}
\details{
The function creates a graph showing Lustgarten's stability measure ASM (N)
for top-N variables N = 5,10,15,20,30,40,50,75,100 for selected FS
The function creates a graph showing the dependence of the selected metrics model(N)
N = 5,10,15,20,30,40,50,75,100 for selected FS
}
\examples{

\dontrun{

decisions <- data$class
data$class <- NULL

result <- ensembleFS(data,
           decisions,
           methods = c('fs.utest', 'fs.mrmr'),
           method = 'kfoldcv',
           params.cv = list(k = 3, iter = 10),
           level.cor = 0.75,
           params = list(adjust = 'fdr', feature.number = 10, alpha = 0.05),
           asm = c('fs.utest', 'fs.mrmr'),
           model = c('fs.utest', 'fs.mrmr')
           )



graph.result(result$stability, 'stability')

graph.result(result$model, 'acc')

}
}
