% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.model.R
\name{model.result.top.var}
\alias{model.result.top.var}
\title{Train model Random Forest for the top-N variables N = 5,10,15,20,30,40,50,75,100}
\usage{
model.result.top.var(x, y, list.selected.var, list.index.cross)
}
\arguments{
\item{x}{input data where columns are variables and rows are observations (all numeric)}

\item{y}{decision variable as a boolean vector of length equal to number of observations}

\item{list.selected.var}{A \code{\link{list}} with selected variables in cross-validation}

\item{list.index.cross}{A \code{\link{list}} with indexes obtained in cross-validation}
}
\value{
A \code{\link{data.frame}} with metrics Accuracy, AUC, MCC for the top-N variables N = 5,10,15,20,30,40,50,75,100
}
\description{
Train model Random Forest for the top-N variables N = 5,10,15,20,30,40,50,75,100
}
\details{
Train model Random Forest in cross-validation and variables selected in the cross-validation
for the top-N variables N = 5,10,15,20,30,40,50,75,100
}
\examples{
\dontrun{

class <- data$class
data$class <- NULL

list.index.cross <- cross.val(x = data,
                             y = decisions,
                             method = 'kfoldcv',
                             params.cv = list(niter = 10, k = 3))

list.selected.var <- feature.selection(x = data,
                             y = class,
                             method = 'fs.utest',
                             list.index.cross = indexes,
                             params = list(adjust = 'holm'))

model.result <- model.result.top.var(x = data,
                                     y = class,
                                     list.selected.var = list.selected.var,
                                     list.index.cross = list.index.cross)

}

}
