% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.ranking.R
\name{ranking.feature}
\alias{ranking.feature}
\title{Function for ranking selected variables}
\usage{
ranking.feature(list.selected.var)
}
\arguments{
\item{list.selected.var}{A \code{\link{list}} with selected variables in cross-validation}
}
\value{
A \code{\link{data.frame}} with the result of the rating of the variables that were most often performed in each iteration of cross-validation
}
\description{
Function for ranking selected variables
}
\details{
Ranking is calculated based on selected variables that are more common in cross-validation iterations.
}
\examples{

\dontrun{

class <- data$class
data$class <- NULL

list.index.cross <- cross.val(x = data,
                             y = decisions,
                             method = 'kfoldcv',
                             params.cv = list(niter = 10, k = 3))


list.selected.var <- feature.selection(x = data,
                             y = class,
                             method = 'fs.utest',
                             list.index.cross = list.index.cross,
                             params = list(adjust = 'holm'))

ranking.var <- ranking.feature(list.selected.var = list.selected.var)

}

}
