% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.selection.R
\name{stability.selection.top.var}
\alias{stability.selection.top.var}
\title{Compute Lustgarten's stability measure ASM (N) dependence for top-N variables N = 5,10,15,20,30,40,50,75,100}
\usage{
stability.selection.top.var(list.selected.var, list.index.cross)
}
\arguments{
\item{list.selected.var}{A \code{\link{list}} with selected variables in cross-validation}

\item{list.index.cross}{A \code{\link{list}} with indexes obtained in cross-validation}
}
\value{
A \code{\link{data.frame}} with the result Lustgarten's stability measure ASM (N) dependence for top-N variables N = 5,10,15,20,30,40,50,75,100
}
\description{
Compute Lustgarten's stability measure ASM (N) dependence for top-N variables N = 5,10,15,20,30,40,50,75,100
}
\details{
Compute Lustgarten's stability measure ASM (N) dependence for top-N variables N = 5,10,15,20,30,40,50,75,100
for variables selected in cross-validation
}
\examples{
\dontrun{

class <- data$class
data$class <- NULL

list.index.cross <- cross.val(x = data,
                             y = decisions,
                             method = 'kfoldcv',
                             params.cv = list(niter = 10, k = 3))

list.selected.var <- feature.selection(x = data,
                             y = class,
                             method = 'fs.utest',
                             list.index.cross = indexes,
                             params = list(adjust = 'holm', alpha = 0.05))

result.stability <- stabilty.selection.top.var(list.selected.var, list.index.cross)

}

}
