% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.selection.R
\name{stabilty.selection}
\alias{stabilty.selection}
\title{Compute Lustgarten’s stability measure}
\usage{
stabilty.selection(list.selected.var, list.index.cross, nvar)
}
\arguments{
\item{list.selected.var}{A \code{\link{list}} with selected variables in cross-validation}

\item{list.index.cross}{A \code{\link{list}} with indexes obtained in cross-validation}

\item{nvar}{the number of first variables for which to calculate}
}
\value{
value stability of selection (Lustgarten’s stability measure)
}
\description{
Compute Lustgarten’s stability measure
}
\details{
Compute Lustgarten’s stability measure for variables selected in cross-validation
}
\examples{

\dontrun{

class <- data$class
data$class <- NULL

list.index.cross <- cross.val(x = data,
                             y = decisions,
                             method = 'kfoldcv',
                             params.cv = list(niter = 10, k = 3))

list.selected.var <- feature.selection(x = data,
                             y = class,
                             method = 'fs.utest',
                             list.index.cross = list.index.cross,
                             params = list(adjust = 'holm', alpha = 0.05))

asm <- stabilty.selection(list.selected.var, list.index.cross, 5)

}

}
