% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.information.R
\name{get.top.gene}
\alias{get.top.gene}
\title{Get the best genes}
\usage{
get.top.gene(list.imp.var.cv, level.freq, number.gene)
}
\arguments{
\item{list.imp.var.cv}{the main genes obtained in \code{ensembleFS}}

\item{level.freq}{how many times a gene has occurred in of the feature subsets}

\item{number.gene}{number of genes}
}
\value{
A \code{\link{list}} with top genes for each method specified in \code{ensembleFS}
}
\description{
Get the best genes
}
\details{
Get the best genes out obtained in \code{ensembleFS}
}
\examples{
\dontrun{

decisions <- data$class
data$class <- NULL

result <- ensembleFS(data,
            decisions,
            methods = c('fs.utest', 'fs.mrmr'),
            method.cv = 'kfoldcv',
            params.cv = list(k = 3, iter = 10),
            level.cor = 0.75,
            params = list(adjust = 'fdr', feature.number = 10, alpha = 0.05),
            asm = c('fs.utest', 'fs.mrmr'),
            model = c('fs.utest', 'fs.mrmr')
            )

gene.top <- get.top.gene(result$selected.feature, 15 , 20)

}

}
